Refactoring: moved utils in internals, moved tests in laws
