+++ core/shared/src/main/scala/Playground.scala
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2017 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * Copyright 2017 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+object Playground {
+  import scala.concurrent.ExecutionContext.Implicits.global
+  import cats.effect.IO
+
+  def sequence(list: List[IO[Int]]): IO[Int] =
+    list.foldLeft(IO(0)) { (acc, io) =>
+      acc.flatMap((sum: Int) => io.map((e: Int) => sum + e)).shift
+    }
+
+  val io = sequence((0 until 100000).map(IO(_)).toList)
+  io.unsafeToFuture().onComplete(println)
+}
