+++ core/shared/src/main/scala/cats/effect/internals/NonFatal.scala
@@ -14,17 +14,22 @@
  * limitations under the License.
  */
 
-package cats.effect
+package cats.effect.internals
 
-import cats.effect.util.TestContext
-import org.scalactic.source
-import org.scalatest.{FunSuite, Matchers, Tag}
-
-private[effect] class BaseTestsSuite extends FunSuite with Matchers {
-  /** For tests that need a usable [[TestContext]] reference. */
-  def testAsync[A](name: String, tags: Tag*)(f: TestContext => Unit)
-    (implicit pos: source.Position): Unit = {
-
-    test(name, tags:_*)(f(TestContext()))(pos)
+/**
+ * Extractor of non-fatal `Throwable` instances.
+ *
+ * Alternative to [[scala.util.control.NonFatal]] that only
+ * considers `VirtualMachineError`s as fatal.
+ *
+ * Inspired by the FS2 implementation.
+ */
+private[effect] object NonFatal {
+  def apply(t: Throwable): Boolean = t match {
+    case _: VirtualMachineError => false
+    case _ => true
   }
+
+  def unapply(t: Throwable): Option[Throwable] =
+    if (apply(t)) Some(t) else None
 }
